Split out additional workspace control panel buttons into components
