+++ __tests__/src/components/WorkspaceControlPanel.test.js
@@ -9,31 +9,11 @@ describe('WorkspaceControlPanel', () => {
   beforeEach(() => {
     store.dispatch(actions.receiveManifest('foo', fixture));
     store.dispatch(actions.receiveManifest('bar', fixture));
-    wrapper = shallow(<WorkspaceControlPanel store={store} />).dive().dive();
+    wrapper = shallow(<WorkspaceControlPanel store={store} />).dive();
   });
 
   it('renders without an error', () => {
     expect(wrapper.find('WithStyles(Drawer)').length).toBe(1);
     expect(wrapper.find('Connect(miradorWithPlugins(WorkspaceControlPanelButtons))').length).toBe(1);
-  });
-
-  it('renders a list item for each manifest in the state', () => {
-    expect(wrapper.find('ul Connect(ManifestListItem)').length).toBe(2);
-  });
-
-  describe('handleAddManifestClick', () => {
-    it('sets the anchor state', () => {
-      wrapper.instance().handleAddManifestClick({ currentTarget: true });
-
-      expect(wrapper.dive().find('WithStyles(Menu)').props().open).toBe(true);
-    });
-  });
-
-  describe('handleAddManifestClose', () => {
-    it('resets the anchor state', () => {
-      wrapper.instance().handleAddManifestClose();
-
-      expect(wrapper.dive().find('WithStyles(Menu)').props().open).toBe(false);
-    });
   });
 });
