+++ src/components/WorkspaceControlPanelButtons.js
@@ -3,6 +3,8 @@ import PropTypes from 'prop-types';
 import List from '@material-ui/core/List';
 import miradorWithPlugins from '../lib/miradorWithPlugins';
 import ConnectedWorkspaceFullScreenButton from './WorkspaceFullScreenButton';
+import ConnectedWorkspaceAddButton from './WorkspaceAddButton';
+import ConnectedWorkspaceMenuButton from './WorkspaceMenuButton';
 /**
  *
  */
@@ -16,6 +18,8 @@ export class WorkspaceControlPanelButtons extends Component {
     const { children } = this.props;
     return (
       <List>
+        <ConnectedWorkspaceAddButton />
+        <ConnectedWorkspaceMenuButton />
         <ConnectedWorkspaceFullScreenButton />
         {children}
       </List>
